添加下划线式可滚动的视图【segmentedControl】

+++ HSYCocoaKit/Classes/Category/UIImageView+UrlString.m
@@ -46,7 +46,7 @@
       placeholderImage:(UIImage *)placeholderImage
 {
     if (error) {
-        NSLog(@"图片请求失败，error = %@, error.code = %ld", error, error.code);
+        NSLog(@"图片请求失败，error = %@, error.code = %ld", error, (long)error.code);
         self.clipsToBounds = YES;
         self.contentMode = UIViewContentModeScaleAspectFill;
         UIImage *realPlaceholderImage = placeholderImage;
